day 11: fix mirrored output
I originally calculated my matrices wrong -- I used the Cartesian plane
(where up is the +y axis), but I wrote the rest of the program assuming
a more computer-graphics plane (where +y is *down*, +x is right).

The basis vectors for my transforms were simply inverted; the
left-column of the "turn left" matrix is what-happens-to-(1,0): it
becomes "up" a.k.a. (0, -1).  And likewise rethink the other three
vectors, which ended up with having exactly inverted left and right.