intcode: give state `impl`s instead of fn params
This is a bit cleaner, since the State can encapsulate all of the
addressing modes rather than having to pass juuuust enough of its pieces
around as function parameters.

The downside is that it's now less obvious which pieces of State are
being mutated by each of the calls, so I named the ones that actually
change the instruction pointer "consume_*" to indicate that.  The others
are mutable merely to resize the memory array when needed.