day 14: iterate through what is needed and extra
This means we make reactions in batches of output_size, and just loop
through adding them to our inventory of "what we have produced", which
is super unoptimized but it passes the tests in debug mode pretty
quickly.

I really don't like the .get_mut()s and the .clone()s in here, but
HashMap apparently does not implement IndexMut.